Split socket configuration from main configuration
This is going to be necessary for future work towards more robust
"reload" mechanisms.  Getting it out of the way now in the main
tree will reduce the size and complexity of those patches.

Mostly what's going on in this patch is the splitting of
socket-related things from cfg_t to socks_cfg_t, the splitting
of socket-related config code from conf.[ch] to socks.[ch], and
then various updates elsewhere to cope with those two things.