deprecate listen => scan
I believe that the current default 'any' option works for all the
use-cases people would have had for 'scan' in the distant past,
and has for a long time now (thanks to ancillary socket data and
SO_REUSEPORT).  In some cases, explicit address configuration
might be a better option than 'any', but I just don't see the
pragmatic case for using the current 'scan' behavior for an
authdns server.  It was at one time the default, but that was
primarily for lack of being able to work around ANY-address
problems via ancillary data.

The motivation here is that the scan code is pretty hacky and
heuristic and causes all sorts of code-factoring problems for
future work because it needs the ability to soft-fail bind() very
late in the startup sequence.

It's entirely possible someone will scream when this change lands
and point out some reason that the option is necessary in a real
use-case that should be supported, in which case I'll have to
either bugfix 'any' to work for them or un-deprecate 'scan'.  I
hope not, but pushing this is the easiest way to find out.