Remove some pointless network compat checks/hacks
This removes several pointless checks from configure.ac for
standard network library defines and functions which basically
never fail to exist on any current, reasonable platforms AFAIK,
and it's not like we had any real compatibility fallback if they
did (either fail autoconf or fail compile, take your pick!).

Perhaps the biggest thing here is dropping the autoconf check for
"-lnsl -lsocket" as required by Solaris (and other SVR4?).  Even
in the Solaris case it doesn't seem like a good idea to bother
trying.  Considering that we do lots of tricky things which
probably break on Solaris variants anyways, blindly dropping in
"-lnsl -lsocket" was just allowing it to maybe-compile with
probably-broken behavior on a mostly-untested platform that seems
to be slowly dying off (outside of Oracle anyways).