regression suite resiliency stuff
Apparently under "prove", Test::More::BAIL_OUT can skip over END
blocks somehow (at least, that's the behavior I seem to be
getting), so replace that with a simple "die" (which is not the
same, but the result is better).

Also, the 001basic checks for specific testsuite modules were
outdated vs the list in _GDT.pm and redundant anyways.