monitored state transitions => syslog
The log output for monitored state transitions was accidentally
left out of the rewrite of the state code for 2.x.  This adds
several new tests explicitly for the new log output, as well as
log-checking tests of the interaction between forced admin states
and plugin-monitored states.  Some further testsuite delay
reduction was done in related code as well.