testsuite: clean up gdnsd_extmon_helper -related path logic
This commit is mostly just to reduce future confusion.

Long ago the installation path of gdnsd_extmon_helper switched
from bindir to libexecdir, but (a) the testsuite was never updated
to reflect that and (b) the testsuite never really needed path
info/fixups for the "make installcheck" case anyways, only for
running a normal "make check" in the build dir (which is why the
wrongness here never really caused an issue).

As a minor posiive side-effect, the testsuite no longer tries to
pointlessly pre-execute gdnsd_extmon_helper in the installcheck
case, which was using the wrong path and then ignoring the error.