require existence of config dir on startup
With an execve reload, if we don't get an absolute configdir path
before the daemon's chdir("/"), we lose the ability to ever find
it again, and without it existing we can't use realpath() either
(and anything less is suspect, really).  This was always a
potential corner-case issue for other existing runtime-reload uses
anyways.  It's better to simply require the directory's existence.