gdgeoip2: libmaxminddb 1.2.0+ compatibility
With this commit, libmaxminddb 1.2.0+ is supported via its newer
database iteration interface.  Prior to this commit, even the our
1.1.5-compatible code from v2.2.3 caused runtime errors with
1.2.0, as the the library-internal assumptions required by the old
interface were broken (again) in 1.2.0.

libmaxminddb 1.2.0 is a runtime compatibility barrier for gdnsd,
in spite of its lack of a major version bump:

If gdnsd is built against 1.2.0 or higher, it will not work
correctly if the runtime library is downgraded to 1.1.x or lower
without rebuilding.  It will exit gracefully with a note about the
library version issue.

If gdnsd is built against 1.1.x or lower, it will not work
correctly if the runtime library is upgraded to 1.2.0+ without
rebuilding gdnsd.  It will exit gracefully with a note about the
library version issue.