gdgeoip2: Refactor offset caching to support offset of zero
The offset cache code made the assumption that an offset of zero
would never occur as legal data to look up, but instead always
signaled defaulting due to empty data.  This removes that
assumption from the cache code (it can now cache for all possible
offset values), and moves the zero/default check down into the
uncached version of the function where it logically belongs (and
where it's easier to remove for 1.2.0).