hashing cleanup: static lookup2 everywhere
We have a central (static, inline-able) implementation of lookup2
that was re-used in various places, and also multiple separate
implementations (with slight interface variations) of the
more-simplistic DJB hash (xor variant).

This rids the code of all DJB hash variants, replacing them all
with short wrappers around gdnsd_lookup2() which are static and
inline-able.  lookup2 can be slightly slower but it's a much
better hash.  In almost call cases we'd prefer its robustness, and
in 2016 the minor CPU perf diff between the two is trivial.

Note: as a special case, gdnsd_dname_hash also retains its library
implementation, so that external plugins don't need a re-compile.
The hacks around gdnsd_dname_hash_static can be removed in a
future major-version bump.