Fixup NONNULL assertions everywhere
This unfortunately-large commit removes redundant dmn_assert(foo)
on function arguments explicitly declared as non-NULL by
function-level attributes.  Long ago the redundancy was desirable
as I wasn't sure about the degree to which various compilers were
enforcing and/or optimizing on the attribute information.  At this
point in history they seem to take it much more seriously and thus
spam the builds with warnings about the redundant checks'
pointlessness.