Stats/States: only JSON and only via csock
This removes from the stats/states code (and support elsewhere) for:
  * CSV output format
  * HTML output format
  * HTTP listener
  * Syslog emitter

The JSON output format is retained, but is only accessible via the
control socket.

"gdnsdctl stats" dumps the current stats in JSON form from the
control socket to stdout.  Periodic file output can be driven by a
cronjob executing gdnsdctl if needs be, or another tool that
attaches directly to the control socket.

"gdnsdctl states" is similar and outputs monitored state
information, in a slightly-updated format from previous
iterations.

With the stats HTTP listener gone, a few remaining optimizations
for later and simpler DNS socket binding are achieved along the
way as well.