dmn: simplify log buffer allocation
Replaces the complex malloc-based scheme for up to 21K of
formatted data with a simpler one that supports up to 4K.

Basically, >4K of formatter strings in a single log message is
already an unreasonably-high limit, and this gets rid of the
internal malloc and potential failure-point there, and simplifies
thread resource cleanup by avoiding the attachment of malloc
results to a TLS pointer.

While here, also update size arguments to size_t, and make all log
formatters explicitly returns_nonnull.