Log the appropriate error in the sendmmsg() case
When sendmmsg() fails with a negative return value, we should be
logging about the true "errno", rather than doing a getsockopt to
retrieve the SO_ERROR value.  I'm not actually sure if the
SO_ERROR method ends up retrieving the same value or not, but this
is more-correct and cheaper.

Not a very serious bug, worst case we were logging the wrong
reasons for send failures, but still counting them in stats and
logging about them at all.  This has apparently been done "wrong"
here since before git history began and nobody's really noticed.