Socket API compatibility updates
All of these seem broadly compatible now on recent *BSD and Linux
releases (recently enough for a major version bump in 2018,
anyways):

* Assume SO_REUSEPORT is available
* Assume accept4() exists
* Assume SOCK_NONBLOCK exists for use with socket() and accept4()
* Assume SOCK_CLOEXEC exists for use with socket() and accept4()
* Allow for detection/use of sendmmsg() interfaces on non-Linux
  (seems to now exist for at least NetBSD and FreeBSD, probably
  other variants too)

It's entirely possible we'll end up backtracking on some of this
after a valid complaint of "Hey gdnsd doesn't build/work anymore
on os A release B which still has years of active life/support",
at which point we can go retrofit the necessary level of backwards
compatibility and document any caveats.  I'm hoping that ends up
not being the case for the most part, and it's a good time to hit
the reset button on such cruft that exists today.