Smooth daemon takeover
This adds a new commandline flag "-T" for server
"start"/"daemonize" actions, which allows it to takeover smoothly
from another running daemon, as opposed to erroring out early due
to noticing the existing server's control socket lock.  Because
the takeover process is between two independent daemon processes,
this form of smooth restart supports both configuration changes
and code upgrades.

Takeover coordination is accomplished via control socket
connection between the old and new daemons.  All DNS sockets are
handed off via SCM_RIGHTS.  This means the old and new DNS sockets
share kernel-level buffers, and thus there is no request loss
during the takeover, but there is a (usually very) brief window of
time when both the old and new daemons are answering queries in
parallel.

In the case that socket/thread counts have changed in the new
configuration (e.g. changes in udp_threads, tcp_threads, and/or
the set of listening addresses), excess sockets not needed by the
new daemon will be discarded, and genuinely-new sockets will be
created where necessary.  In the case that the socket/thread count
for an address is reduced by the new configuration (e.g. listen
addresses stay the same, but udp_threads reduced from 8 to 4),
there will unavoidably be some request packets lost in the receive
buffers of the sockets being permanently closed off during the
takeover operation.

The control socket itself, and its locked lockfile, are also
handed off across the takeover control socket connection, which
avoids race conditions on re-establishing these.  The control
socket never stops listening for new clients, but does stop
actively calling accept() to process them briefly during the
handoff, to ensure that successful control socket connections stop
being received by the old daemon before starting reception at the
new.

Only one takeover attempt is allowed concurrently by the existing,
established daemon.  Other takeover requests will be denied until
the first takeover request either succeeds (resulting in the death
of the old daemon), or fails (by the old daemon detecting that the
new daemon died before it completed the process).  On takeover
failure, all relevant state is re-set in the original daemon so
that it behaves as if the takeover attempt never happened.  While
a takeover attempt is in progress, the existing daemon also denies
stop requests.

Building on top of all the above, there's also a new gdnsdctl
command "replace".  This allows gdnsdctl to request that a daemon
replace itself by using fork()->fork()->exec() to spawn a new
orphaned daemon to takeover from it.  This allows retaining
certain attributes of the original daemon's execution context
which preserve across fork and exec boundaries, including,
critically, Linux cgroups for systemd service compatibility.
Other attributes preserved may aid in writing init scripts /
configurations for other init systemd as well (vs always having to
launch the takeover daemon independently).

For systemd, this is all integrated by setting ExecReload= to
"gdnsdctl -l replace".

The "replace" command is synchronous and status-reporting in the
same way as "stop".   The command will wait for success or failure
of the replace transaction with the old server (which encompasses
the entire spawning and handoff process), wait for the old server
to exit, and connect to the control socket of the new server for
confirmation, before exiting with an exit status code of zero for
success and non-zero for failure.