ztree: add missing update locks around two assignments
There were two cases here where the old value is NULL and a
newly-allocated pointer is assigned without using the "lock" parts
of the API.  These don't need synchronize_rcu() in the urcu case,
so it doesn't really matter in that case at all in practice.

In cases where the build lacks urcu and falls back to pthread
rwlocks, there are more correctness concerns here.  If you spend
some time thinking through all the logical constraints and
consequences, in practice it still doesn't actually cause any
correctness issues.

But it took me a while to even figure out that it wasn't an issue,
and I'd have to add a few paragraphs of commentary prose just to
explain it to my future self or anyone else, so it's probably
better to just put the locks in place and be done with it :)