C Code formatting
Code formatting commits are annoying because they cause a giant
pain point in the history around which rebases and merges are
difficult.  Now is an opportune time to do it with so much else in
flux.

My coding style has evolved over the years, but I've generally
stuck to my years-old style for self-consistency here.  I've also
been inconsistent about style issues that matter less to me.  The
real sin here, though, is this repo not having a defined coding
style with a check/correct tool at all, which has allowed lots of
other minor issues to creep in.

With this commit, I've made some arbitrary-but-reasonable choices
that can be forced with "astyle" (using version 2.06), and added
some qa scripts for checking/correcting, applied astyle's fixups,
and then manually corrected the places where the fixes looked ugly
and/or wrong to the point where both astyle and I are
reasonably-happy with the final output.