getsockopt socklen_t fixup in 2 places
socklen_t* is the correct type to be passing in the final argument
of getsockopt(), as in other examples in the codebase.  This is
mostly about being pedantic, though, as socklen_t will almost
universally be equivalent to "int" for hysterical raisins, which
is the same width as our "unsigned" was here.