Bugfix: geoip2: do not treat missing coords as 0,0
In the specific case of GeoIP2 databases and auto_dc_coords, when
encountering a coordinate-free network entry (e.g. satellite,
anon-proxy), the lookup will now return the configured default
datacenter ordering.

This is the documented, correct behavior and matches what we've
always done in the GeoIP1 case as well.  Before this commit, the
code was resolving such networks from GeoIP2 databases as if they
were literally located at 0,0 and picking a datacenter list based
on distances from that point.