Warn in docs and code about 2019 GeoLite2 changes
All GeoIP2 databases will have some networks mapped to objects
that do not contain coordinates (satellite, anon-proxy, and some
unlabeled unknowns as well), and plugin map statements can overlay
fixed and coordinate-based mapping to varying degrees, so there's
not even a reliable measurement we could use like "X% of
coordinate lookups must succeed" to try to figure this out
precisely and automatically.

So, while it would have been nicer to detect this based on the
data, it's not realistic in terms of complexity, cost, or
accuracy.  Our only reasonable alternative here is to rely on the
database's own metadata description abd build epoch times.

Fixes #161