UDP: always use fixed mmsg=16 if avail, remove VLAs
Before this commit, mmsg width was configurable, and could be set
to 1 to disable the use of mmsg interfaces altogether.

The configuration option 'udp_recv_width' was removed, and the
code now always uses mmsg interfaces with a fixed with of 16
(double previous default) if compile- and run- time checks are
successful.  This removes the last complex set of VLA uses.

Along the way various improvements were made to how the
mmsg/cmsg-related buffers and structures are allocated and
initialized which should increase efficiency and locality, as well
as how skipped (no response) packets are dealt with inbetween
recvmmsg() and sendmmsg().

The basic rationale for the fixed width: with a fixed value of 16,
we already have a bit over an order of magnitude improvement over
the baseline of single-packet recvmsg().  This should be plenty
from the perspectives of absorbing microbursts from the buffer and
increasing the efficiency of the userland<->kernel barrier for the
socket operations.

At much larger values, there's the potential for this to act in a
bufferbloat-like way and increase latencies, especially since the
inner packet processing work is serialized for a given I/O thread.
To get beyond these limits in terms of throughput, the right
answers aren't huge widths here, but rather one of the following
from various perspectives: more udp_thread count, improving kernel
socket API efficiency in general, hardware upgrades, or just
letting the packets drop.