plugin_weighted: remove VLA usage in resolve()
This removes the (potentially-large) VLA "dyn_addr_weights[][]"
from resolve(), replacing it with a pointer to a TLS-allocated
malloc region initialized at iothread startup.  Note the use of
pointer-to-2D-array syntax, which is a little bit tricky but is
standards-conformant, avoids true VLA usage, and allows the code
to still reference it using dyn_addr_weights[][] syntax.