Include ragel outputs in dist tarballs
I'm backtracking a bit here, but those that build from source for
every install (hi Gentoo and BSD Ports users!) complain a lot
about the dependency hell dragged in by needing ragel as a
dependency to build/install gdnsd.

This commit puts things back the way they used to be: the ragel
output still isn't stored in the git repo, but is included in the
distribution tarball, and thus ragel won't be necessary so long as
timestamps work sanely and you don't try to edit the ragel input
files ending in ".rl".

Fixes gdnsd/gdnsd#133