Use /tmp for testsuite rundir paths
Now that we rely on a control socket for everything, the traditional
length limits on control socket pathnames (108 on Linux!) are becoming
annoying.  The testsuite often gets run from deep in directory
hierarchies, which makes the problem much easier to trigger (vs real
running installations).

My example annoyance: if I start with a git checkout in
"/home/bblack/repos/gdnsd/" and run "make distcheck", several test
scripts fail for overlong daemon control socket paths with lovely
examples like this 112 byte one:

/home/bblack/repos/gdnsd/gdnsd-2.99.1/_build/sub/t/testout/022anytcponly_051_any_tcp_only/run/gdnsd/control.sock

Latter parts of that path could be minimized (somewhat), but there's no
escaping the fact that the testsuite shouldn't inexplicably fail
depending on how long the pathname was to your build directory.

Hopefully the fixed use of absolute "/tmp" won't cause issues for
others.  There is a hook at the end of the testsuite that tries to wipe
out all the directories created by this.

Another alternative, if /tmp proves problematic, would be to use
relative pathnames and have the testsuite chdir() to make them work, but
that also sounds pretty fraught with strange edge cases.