Add support for EDNS0 NSID (RFC 5001)
This adds a new configuration option "nsid", which takes an ascii
hexidecimal string up to 256 characters (128 encoded bytes) long.  If
and only if the option is specified, all responses to EDNS0-capable
clients will include this data as the NSID option.  Useful for
non-transitive identification of servers in any anycast group or behind
a loadbalancer.  The meaning of the hex bytes is arbitrary and up to the
administrator (see the RFC!).