Automatic serial from mtime for rfc1035 zonefiles
If the serial field of an SOA record is set to zero, an automatic
serial number based on the zonefile's last modification timestamp
will be generated with 1-hour resolution using the format
YYMMDDHHMM in the UTC timezone.  This gives 1-minute resolution
and will fail in the year 2043, by which time we'll have to come
up with something new.  If there were $INCLUDE files involved, the
effective last modification will be the latest of all recursively
included files.