gdnsdctl: allow acme challenges with leading dash
The base64url strings used in acme-dns-01 challenges on the
commandline can legitimately have a leading '-' character, which
GNU libc's getopt() in its default mode will interpret as an
option, because it scans the whole of argv and permutes
non-options to the end of the list while slurping all options.
POSIX behavior is to not permute and stop at the first non-option,
which avoids the issue.