Cleanup memory in non-debug builds
I got tired of explaining these cases to various analysis tools,
where we skipped various destruction steps when the whole daemon
was exiting anyways in non-debug builds.  It's not that costly to
just free it all even in production builds and wipe out some cruft
and confusion.

This whole commit is fairly mechanical.  It just renames the
"debug" part out of various cleanup function names and related
defines, and gets rid of the various "#ifndef NDEBUG" parts
related to them as well.