Revert "dlclose() plugins on shutdown"
This reverts commit 0083d8f5825714bb5a7c609ee19a2216e07f1cc4.

This turns out to be unwise paranoia.  It's not going to help in
the cases I was paranoid about because the dlclose() is only going
to happen in the original exiting process anyways, not in the
middle process or the one running execlp().  Further, this causes
address sanitizers to call all data indirectly referenced from
plugins' addresses spaces to be indirectly leaked on shutdown,
which is a PITA to fix.