Restore CNAME chain output, restrict DYNC usage
This is an unfortunately-deep change this late in the 3.x
test->release cycle, sorry!

TL;DR:
* Zone-local CNAMEs are (once again, like 2.x) followed when
  generating response packets.
* Dynamic CNAME target values emitted via DYNC RRs can no longer
  point into same-zone data.

During interop testing of 2.99.x, I ran a bunch of tests through
thousands of random public open recursors and found evidence that
a very small handful of them were in fact not compatible with our
earlier change to not chase zone-local CNAME targets during
response generation.  This was known and noted to be the most
standards-dangerous of 3.x's extreme minimization efforts.  It
turned out to be a bridge too far!

The few recursors that caused issues could eventually be coaxed
into emitting correct answers, but only after re-sending them the
query multiple times (adding one CNAME chain entry to their cache
set for each attempt until they finally construct it all).  I
think they were ancient versions of BIND and HyperDNS, but it's
hard to be sure.

This commit reverses our course and restores full standards
compatibility in this area: CNAME chains within a zone's data will
be chased and emitted until they reach either (a) a local name
that lacks CNAME, (b) a delegation cut, or (c) a CNAME that points
into another zone entirely.  As before in 2.x, there is still a
hard depth limit of 16 when recursing CNAME chains in zone-local
data, and as in earlier 2.99.x releases, this is no longer
configurable to any other value.

One of the reasons for even trying to curb CNAME chain outputs was
that it removed one of the final barriers to easily verifying the
limits on total output packet size of all possible responses at
the time of zone loading (and thus eventually pre-encoding most
response data, sometime in the future), and this in turn had
positive implications elsewhere for simplifying compression, etc.

In order to restore CNAME chain outputs and still realize most of
these benefits, something else had to give, and that something
else was to impose new restrictions on DYNC instead:

* DYNC plugin configs now require that their CNAME response values
  be configured as fully-qualified names ending in dots (we
  could've just called them all fully-qualified with or without
  trailing dots, but then it would've made the breakage silent for
  users whose configurations currently rely on the now-unsupported
  relative names).

* DYNC can no longer be used to emit CNAMEs that target zone-local
  data.  When zone data is loaded, a check is performed on DYNC
  records that can see into all their possible response values,
  and the zone load will fail if any of them point to the same
  zone as the DYNC RR itself.

Pointing DYNC CNAME data at external names only is more in line
with the original design intent of DYNC anyways, but this makes it
a real restriction, and some users may find their configurations
have to be refactored around this.  The upside for the code is
that these restrictions imply that CNAME output by a DYNC doesn't
recurse back into local data, and thus removes DYNC's problematic
runtime unpredictability in three key areas: its impact on final
output packet sizes, whether it would cause a CNAME loop, and its
effects on compression.