Bugfix: move ltree zwarns out of p1_rrset_size
This is a bugfix for the CNAME-related work c9fc7de

p1_rrset_size() does not return a boolean failure status like the
rest of the phase1 functions (it returns a size_t packet size),
and therefore log_zwarn() statements shouldn't be used there, as
they can return boolean failure status if "zones_strict_data" is
turned on in the configuration or via the commandline option -S.

The net effect of the bug was that in the case "zones_strict_data"
was true, and there were MX/SRV warnings to issue (missing target
addresses) which the strictness upgrades to fatal errors, the
error would not be fatal, and the rrset sizing check would
effectively be skipped (true is 1 as a size_t), which could have
allowed illegaly-large data to be configured successfully.

While here, also fixed a related bug that MX/SRV target address
warnings could be emitted multiple times (once for each start
point of a CNAME chain leading into them, in addition to the
warning when actually checking them directly).

---

This is the kind of bug that -Wconversion might have pointed out
to us, I think?  I'm still on the fence about moving the code to
Wconversion cleanliness because it can add a lot of line noise in
pointless places and tends to push you in the direction of
size_t-ification.  I'm not even sure where that rabbithole ends
at, although simply going through the process of fixing them
manually at least once and then tossing out the changes might at
least shed some light on whether any other existing implicit
conversions are actually bugs.