Simplify ltarena code, kill de-duping hashtable
I don't think the hashtable for dname de-duplication was doing
much anymore; it probably had more utility back in the era when
all zones shared storage.  It may help in some scenarios, but in
most average-case zones it's just a waste.  While at it, moved the
interfaces around a bit, instrumented with better attributes for
bounds checking, cleaned up a few other odds and ends.