systemd unit file: manage rundir
I hadn't noticed this problem on hosts I had upgraded so far,
because they already had a rundir created by gdnsd-2.x and hadn't
been rebooted since the upgrade.  Under systemd /run/ is empty on
reboot, and with User=gdnsd the daemon doesn't have permissions to
create /run/gdnsd for itself, so startup fails.  The correct
answer is adding RuntimeDirectory= and RuntimeDirectoryMode= to
the unit file, which are supported from systemd rev 211.

We probably *should* go a little further in this area and honor
$XDG_RUNTIME_DIR at runtime in the daemon code, to have some
chance of this working in other scenarios without having to
explicitly reconfigure the "run_dir" in the config file, but that
can wait for another day.