Re-arrange shutdown sequence slightly
This moves the ev_signal_stop() calls for the terminating signals
(TERM and INT) much later in the shutdown sequence, just before
our final exit() or raise() call.  This is safer, as otherwise a
fresh SIGINT or SIGTERM could interrupt the orderly shutdown of
threads and child processes, which could unnecessarily fail an
in-progress DNS transaction or leave child process unmanaged,
respectively.