ztree: allow empty set of zones to succeed
This was preventing gdnsd startup with no zonefiles, which would
be the common case for a fresh installation.  It was originally
meant to be a defensive measure against unknown causes that might
make all zonefiles dissappear, but on closer inspection is no
reasonable form of this check wouldn't cause aggravation for at
least some cases.

Don't delete all your zonefiles and then ask for a reload if you
don't want it to happen! :)