control socket response refactoring
Prior to this commit, the only negative response was RESP_NAK,
which gave no hint as to whether there was a real failure or
merely a blocking reason (replacement) to delay/retry.  Also, some
stateful commands like reload-zones and acme-dns-01 ignored the
status of in-progress replacements, which made them a bit
state-lossy during concurrent replacement operations.

With this change, RESP_NAK is split info RESP_FAIL for true
failures (e.g. zone reload failed due to bad data) and RESP_LATR
meaning something more like "can't do this right now because a
replace is ongoing, try again later [on a new connection]".  The
reload-zones and acme-dns-01 commands were made to be
replace-aware, giving RESP_LATR responses during replaces.

This protects against losing the effects of a supposedly
successful reload-zones running concurrently with gdnsdctl
replace, but there's still a lossy window here for independently
spawned takeover daemons, to be dealt with in a later commit.

gdnsdctl is unchanged so far and continues to treat all responses
which aren't RESP_ACK as generic failures.

Relatedly, a third new negative response type RESP_UNK is used
only for unknown request types, in the interest of easier future
compatibility.