Enhance the daemon replacement sequence
This fixes the state-lossy situations mentioned in an earlier
commit by introducing a new control socket exchange during the
takeover process.  REQ_TAK1 is sent by the new daemon very early,
before it has begun loading zonefiles, and provides the necessary
hook for the old daemon to begin denying concurrent reload-zones
requests (and others) with RESP_LATR, which gives the same basic
gaurantees to independent replacements that gdnsdctl-driven
replacements already had.

Note this required a version bump for compatibility checking,
which will allow daemons with this commit to still successfully
takeover from older 2.99.x instances (but not the other way
around, seems pointless for a beta series).  We'll remove this
compatibility later (perhaps 3.1.0 or later, giving a chance to
upgrade from any beta installs to 3.0.x first).

Note also that the previous release "2.99.1729" is actually
"2.9.193" for control socket purposes, because each version field
is converted to a uint8_t with a max value of 255.  For continuity
we'll keep going with values that are actually >= 1729 but which
also roll over to protocol-level versions that increase as well.
The version bump here to .1992 is equivalent to .200