Handle comms errors better during takeover
If the old daemon witnesses a communications error or disconnect
on the new daemon's control socket connection before the final
point of the handoff where the new sends a stop request to the
old, the old now explictly sends the new a SIGKILL to attempt a
recoverable cleanup, in case for some reason the communications
error wasn't going to be followed quickly by natural death of the
new daemon (e.g. new daemon has a bug preventing its exit).