Soft switch to C11 in tooling
This removes the explicit check for C99 and goes with what seems to be
the direction of eventual future autoconf releases in using
AC_PROG_CC_STDC to get the latest-available standards support.  Even
with older autoconf, this seems to avoid explicit -std=c99 if the
default standards version is >= C99.

This should allows us to conditionally use C11 features, if we check
__STDC_VERSION__ for it.

Also, cppcheck updated to use C11 rather than C99 standards.