acme-dns-01: send LATR after data recv
Sending the RESP_LATR response (in case of ongoing replace)
immediately in response to the request doesn't really leave the
connection in a useful state on the server side, as the client
will continue to send challenge data anyways and only looks for
the response afterwards.  This change re-arranges things so that
we always consume the data the client sends (when within
legitimate bounds) first, and then respond with one of RESP_LATR,
RESP_FAIL, or RESP_ACK as appropriate.