csc transaction retvals report retry-ability
Again, nothing uses this yet and existing overall behaviors
preserved, but this allows for gdnsdctl to be reworked in a
futher commit to retry in appropriate cases.  Most send/recv
errors are marked as soft (retryable on a new connection), as they
could represent hitting a close-on-shutdown of a daemon being
replaced.  Explicit RESP_LATR are all marked as soft/retryable.