csc_wait_stopping_server: remove kill/sleep loop
It was a nice idea to do the kill/sleep loop here to confirm
process exit, but I think it was probably a bad idea now because:

1) It's inherently racy.  There's no portable and sane way to
wait on the death of a non-child process in a non-racy fashion.
With this hacky kill/sleep loop, it's entirely possible that the
process exited before we started our loop or during any of the
sleeps in our loop and a new process with the same PID was
allocated before we checked kill() again.

2) It wastes time: pointless time is spent in the kill/sleep loop
in various scenarios.

3) It's ultimately trying to predict an unpredictable, serious bug
in a fairly straightforward part of the code, or with the kernel
or C library.  In all wait_stopping_server() cases, the daemon
being waited on explicitly does not close the file descriptor
before exiting, leaving it to the operating system to close it as
part of its own exit cleanup.  We continue to watch for this
close-on-exit to happen, but the whole pointless point of the
kill/sleep loop was to then witness the virtually guaranteed exit
of the PID fractions of a second later.

4) There's no sane way to tune the timeout between kills, or the
total timeout before giving up.  How long should "reasonable"
process exit sequences take (from close-fds -> PID-vanishes), even
under load?  If some unpredictable bug happens, how long should we
wait before we declare this unknown bug happened?

It seems simpler to assume the basics are working right and that
so long as the operating system closed the socket fd as part of an
exit sequence, there's no need to actually pay attention to the
daemon's PID vanishing during further cleanup.

While here, also remove the log_err() outputs from the function,
letting the caller report as appropriate.  It doesn't much matter
exactly why the recv failed to report close as expected, only that
it did so.