css: remove max_clients code
The max_clients code here turns the accept watcher on and off, but
this is problematic because it causes a denial of service in the
case of certain bugs, and because it conflicts with the replace
sequence's management of the accept watcher.  Since permissions
control access to the socket anyways, we probably don't need this
layer of protection against runaway client connection counts.

We might have to revisit this if in the future we allow the
control socket to be world-writeable and use SCM_CREDENTIALS or
similar to authorize various levels of interaction, but if so we'd
have to do it in a more robust and complex manner than this
anyways.