Add example LSB-style initscript
This renames the "sysd/" directory of the source tree to "init/",
and adds a templated example LSB-style initscript for Linux there
which tries to mirror the functionality of the systemd unit file
to a reasonable degree (doesn't currently cover the filesystem
restrictions).

Also, fixed up the SecureBits setting in the systemd unit file
example (keep-caps is implied by systemd and then pointless due to
no-setuid-fixup, and both the regular and -locked variants of the
other two options should be specified).