Cleanup on numeric constants
In a few places "static const" definitions of numeric constants
were used as if that had its C++ meaning.  These are now switched
to proper C constants using "#define", and should allow better
compile-time constraints and optimizations.

Along the way, cleaned up the definition of some network-order
constants in dnswire.h to use htonl() (which evaluates as
compile-time constant anyways, at least commonly with an
optimizing compiler), which eliminated the last need for the
autoconf check for endianness.