Updates on CAP_NET_BIND_SERVICE matters
Tested and validated that, yes, we can use setpriv and filesystem
permissions from a sysvinit script to run gdnsd securely with
CAP_NET_BIND_SERVICE, if the target system lacks the requisite
kernel 4.3 and/or util-linux 2.31 to support the better pathway of
ambient capabilities, and documented it all more-explicitly for
this case.

Also tested and validated that no version of systemd can do the
same thing, on any kernel, so systemd installs require a 4.3 or
higher kernel and systemd >= v229.

Systemd caused all the mess that led us down this pathway anyways,
so it seems only appropriate and fair that systemd suffers a bit
on the compatibility front here as well :P