Fully parse query packets, answer more robustly
In the past, we've taken convenient shortcuts to only parse for
our common/expected real inputs on today's Internet, and for many
odd inputs we returned variously FORMERR, NOTIMP, or simply
dropped the request with no response.

Some of our logic has always been "wrong" here in ways that matter
little in present practice, but are important for future proofing
gdnsd and not holding back forward progress on standards.

One key example that violates one of the proposed rules of
draft-ietf-dnsop-no-response-issue, is that (prior to this commit)
gdnsd would ignore (no response) packets that lacked query RRs
(all counts zero, just the 12 byte header), and this happened
before other logic like the returning of NOTIMP for unknown
opcodes.

Another key issue is that (prior to this commit) gdnsd could only
successfully find the OPT RR signalling client edns0 support if
the client's query packet had exactly one question, no answer or
auth section records, and the OPT RR was the very first one in the
additional section.  While this arrangement is by far the most
common today, it really hampers our ability to signal edns0
compliance in this face of future extensions to the RRs sent by
clients.

This commit intends to fix all related issues as best we can with
much better query parsing, and perhaps most importantly, this
commit ensures we send some kind of error response packet in
almost every case, leaving very few cases where we fail to send a
response at all.

As of this change, the input query packet parsing rules are as
follows in broad strokes:

* Our maximum receive size in UDP data length terms (as advertised
  in edns0 and enforced on wire reception) is 1024.

* UDP packets longer than this will be truncated (and probably
  parse erroneously, resulting in a FORMERR), and TCP connections
  are immediately closed with a recvfail if the received message
  size indicates >1024 bytes.  We may arguably want to later raise
  this limit in the TCP case for future-proofing, but it hasn't
  been done in this commit.  For UDP I don't see a good reason to
  raise further at this time.  Note our edns0 input size
  advertisement to clients in responses is also 1024.

* The only cases where we ignore input packets (no response) are:
** UDP requests with source port zero
** Requests with less than 12 data bytes (incomplete DNS header)
** Requests with the QR (Query Response) bit set, to avoid loops

* All other queries will get some kind of valid response packet,
  which may have an error like FORMERR

* All of the query packet's RRs as enumerated in the header's RR
  counts (if any) are parsed first before looking at other issues
  like OPCODE and QDCOUNT later.

* If there's at least one question (QDCOUNT) and it can be parsed
  successfully, this question will be reflected in the response
  packet, regardless of other issues/errors.  If there's more than
  one question, the remainder get minimal parsing for correctness
  and to iterate past them, but they are not reflected in any
  response, or used as input.  Zero questions is also legal at
  this point in parsing, but all cases where QDCOUNT != 1 will
  result in some kind of error response much later below.

* The first question cannot have compression pointers in the query
  name (FORMERR).  Others are allowed to, and we don't really care
  where they point at since we're not parsing, using, or copying
  the names.

* Any RRs sent in the Answer, Auth, and Additional sections will
  be parsed, but never reflected (all response data in these
  sections is generated by gdnsd if present at all).  Since we
  don't actually have a need for any this data other than OPT RR,
  we only do the bare minimum parsing to skip through them
  correctly while looking for the OPT RR (type and class is
  ignored and rdlen is used to skip over data).

* The OPT RR will be parsed and used from anywhere within the
  Additional section, assuming the code makes it that far without
  encountering errors while parsing other intervening RRs above.
  Obviously, if RRs before OPT fail to parse, we cannot detect the
  OPT record reliably and will fail to signal edns0.

* If the OPT RR is parsed, the response will always contain our
  matching OPT RR in the output to signal edns0 support, even if
  some unrelated error occurs below.

* If a duplicate OPT RR is encountered, this is a FORMERR (but if
  the first was legal, we will still signal edns0 compliance with
  this FORMERR).

* If there's leftover trailing junk data in the query packet after
  successfully parsing all RRs, this is a FORMERR.

* If we get through all of the query RR parsing above without
  encountering an error, the we look at OPCODE, QDCOUNT, and query
  type/class/name issues in this order:
** If OPCODE is anything other than QUERY: NOTIMP
** If QDCOUNT != 1: FORMERR
** If qtype is a meta-type (128-255) other than ANY: NOTIMP
** If qclass is anything but IN, CH, or ANY: REFUSED
** If qname is not in a zone we're authoritative for: REFUSED